requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
requests>=2.31.0
librouteros>=3.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0 